+++ example/lib/pages/tile_page.dart
@@ -2,13 +2,13 @@ import 'package:flutter/material.dart';
 import 'package:yaru_icons/yaru_icons.dart';
 import 'package:yaru_widgets/yaru_widgets.dart';
 
-class TileList extends StatelessWidget {
-  const TileList({super.key});
+class TilePage extends StatelessWidget {
+  const TilePage({super.key});
 
   @override
   Widget build(BuildContext context) {
     return ListView.builder(
-      shrinkWrap: true,
+      padding: const EdgeInsets.all(20),
       itemBuilder: (context, index) => const YaruTile(
         title: Text('Title'),
         trailing: Icon(YaruIcons.information),
